requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.5.0
opencv-python==4.10.0.84
pillow==10.3.0